Rename segment_buffer file into the new segment_sink name
We changed the name of classes from *SegmentBuffer to *SegmentSink but
didn't actually update the file names. This is now done.
