+++ src/core/segment_sinks/implementations/audio_video/audio_video_segment_sink.ts
