+++ src/core/segment_sinks/implementations/text/index.ts
@@ -1,5 +1,5 @@
-import type { ITextDisplayerInterface } from "./text_segment_buffer";
-import TextSegmentSink from "./text_segment_buffer";
+import type { ITextDisplayerInterface } from "./text_segment_sink";
+import TextSegmentSink from "./text_segment_sink";
 
 export default TextSegmentSink;
 export type { ITextDisplayerInterface };
