+++ src/core/segment_sinks/implementations/text/text_segment_sink.ts
